FileDatabaseInsertionCSV bug fixed in delete table

+++ scripts/FileUploadService/FileDatabaseInsertionCSV.py
@@ -237,8 +237,9 @@ def csv_uploader(parms, dataset_name, user_id=None, tenant=None):
         except Exception, err:
             print err
             result = comm.format_response(False, err, "Error occurred while DataCasting.. \n" + str(err), exception=sys.exc_info())
-            table_delete_status = bq.delete_table(dataset_name,table_name)
-            print 'Table delete status: ' + str(table_delete_status)
+            if parms.folder_type.lower() == 'new' or parms.folder_type.lower() == 'singlefile':
+                table_delete_status = bq.delete_table(dataset_name,table_name)
+                print 'Table delete status: ' + str(table_delete_status)
             return result
         print 'Data casting successful'
         # try:
@@ -276,8 +277,9 @@ def csv_uploader(parms, dataset_name, user_id=None, tenant=None):
         except Exception, err:
             print err
             result = comm.format_response(False, err, "Error occurred while inserting.. \n"+ str(err), exception=sys.exc_info())
-            table_delete_status = bq.delete_table(dataset_name,table_name)
-            print 'Table delete status: ' + str(table_delete_status)
+            if parms.folder_type.lower() == 'new' or parms.folder_type.lower() == 'singlefile':
+                table_delete_status = bq.delete_table(dataset_name,table_name)
+                print 'Table delete status: ' + str(table_delete_status)
             return result
 
         print result
